examples/attack_demo/requirements.txt
canari
openai>=1.0.0
rich>=13.7.0
python-dotenv>=1.0.0
